chore: move client interfaces to origin-backend-core (to avoid cyclic dependencies)
