+++ packages/origin-backend-core/src/index.ts
@@ -11,3 +11,4 @@ export * from './Certificate';
 export * from './LoggedInUser';
 export * from './UserRegistrationData';
 export * from './EmailConfirmation';
+export * from './client';
